% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.sonar.calibration.R
\name{apply.sonar.calibration}
\alias{apply.sonar.calibration}
\title{generate.sonar.calibration}
\usage{
apply.sonar.calibration(
  raw.file = "C:/Users/cbroeckl/Documents/temp/hybridSonar.raw",
  sonar.cal = my.cal,
  recursive = FALSE
)
}
\arguments{
\item{raw.file}{full path (or path within working directory) to the .RAW file to which the calibration is to be applied  An entire directory can be input instead, and if a directory is chosen all appropriate .raw files in that directory will be calibrated.}

\item{sonar.cal}{either an R object generated by 'generate.sonar.calibration, or a path to a csv calibration file written by generate.sonar.calibration.}

\item{recursive}{logical.  if 'raw.file' is a path to a directory of raw files, should the function also look for raw files in all subdirectories? see ?list.files}
}
\value{
returns nothing to the R environment

selected .RAW files will have been modified with a new _sonar.INF and optionally some file copying and renaming if using hybrid sonar.
}
\description{
Create calibration parameters from a low collision energy SONAR scan - traditional or 'hybrid' Sonar
}
\details{
This function uses Apex3D64.exe, distributed by Waters Corp, to perform 4d peak detection in SONAR data, reads that output data, filters, if desired, and calibrates the linear relationships between sonar drift bin and m/z value.  An R-object is returned, and a .csv file exported.  These data contain details on the instrumentation and acquisition method used to collect the calibration data, to ensure that the calibration is only applied to compatible data files.  Masslynx Version, Instrument Serial number, and sonar settings must all match.
}
\author{
Corey Broeckling
}
\concept{Waters SONAR}
\concept{data-independent MS/MS}
\concept{mass spectrometry}
