% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.sonar.calibration.R
\name{generate.sonar.calibration}
\alias{generate.sonar.calibration}
\title{generate.sonar.calibration}
\usage{
generate.sonar.calibration(
 
    raw.file = "R:/RSTOR-PMF/Projects/Broeckling_Corey/SONAR/20220124-hemp-cal/20210805_MSIMM_LT_1080.PRO/Data/20220124_sonarcaltest_016.raw",
  apex3d.dir = "C:/Users/cbroeckl/Documents/SONAR/lib",
  low.energy.function = 2,
  rt.range = c(2, 12),
  cal.out = "cal.out.csv",
  filter = TRUE
)
}
\arguments{
\item{raw.file}{full path (or path within working directory) to the .RAW file from which the calibration is to be made}

\item{apex3d.dir}{full path to the directory containing Apex3d64.exe executable file.}

\item{low.energy.function}{integer value. The mass lynx function containing the low collision energy sonar data. For traditional sonar, set to 1.  For hybrid sonar, set to 2.}

\item{rt.range}{vector of length two.  i.e. c(2,8).  Retention time range, in minutes, for calibration data to be generated on.}

\item{cal.out}{numeric.  the ratio of isotope signal (all isotopes) divided by total spectrum signal * 100 much be greated than minPercentSignal to evaluate charge state. Value should be between 0 and 100.}

\item{filter}{file name or path to export CSV file to.  If you have an existing file and want to append the data, enter full file path/name to that file, and new data will be appended to the same csv file.}
}
\value{
returns an R object which is a data from containing calibration information.

Additionally exports a .csv file with name defined using the 'cal.out' option.
}
\description{
Create calibration parameters from a low collision energy SONAR scan - traditional or 'hyrbid' Sonar
}
\details{
This function uses Apex3D64.exe, distributed by Waters Corp, to perform 4d peak detection in SONAR data, reads that output data, filters, if desired, and calibrates the linear relationships between sonar drift bin and m/z value.  An R-object is returned, and a .csv file exported.  These data contain details on the instrumentation and acquisition method used to collect the calibration data, to ensure that the calibration is only applied to compatible data files.  Masslynx Version, Instrument Serial number, and sonar settings must all match.
}
\author{
Corey Broeckling
}
\concept{Waters SONAR}
\concept{data-indepdenent MS/MS}
\concept{mass spectrometry}
